requirements.txt
beautifulsoup4==4.12.2
celery==5.3.5
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-paginate==2022.1.8
flask-smorest==0.42.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
openai==1.2.2
pinecone-client==2.2.4
python-dotenv==1.0.0
psycopg2==2.9.7
requests==2.31.0
SQLAlchemy==2.0.21
